requirements.txt
awscrt==0.21.1
awsiotsdk==1.22.0
cbor==1.0.0
cbor2==5.4.6
mock==4.0.3
psutil==7.0.0
pytest==8.4.1
setuptools==75.1.0
